support map

+++ src/map.c
@@ -257,7 +257,7 @@ _pbcM_sp_new(int max , struct heap *h)
 	}
 	ret->cap = cap;
 	ret->size = 0;
-	ret->slot = (struct _pbcM_sp_slot *)HMALLOC(ret->cap * sizeof(struct _pbcM_sp_slot));
+	ret->slot = (struct _pbcM_sp_slot *)HMALLOC((int)(ret->cap * sizeof(struct _pbcM_sp_slot)));
 	memset(ret->slot,0,sizeof(struct _pbcM_sp_slot) * ret->cap);
 	ret->heap = h;
 	return ret;
@@ -285,7 +285,7 @@ _pbcM_sp_insert_hash(struct map_sp *map, const char *key, size_t hash_full, void
 			slot->pointer = value;
 			slot->hash = hash_full;
 		} else {
-			int empty = (hash + 1) & (map->cap-1);
+			int empty = (int)((hash + 1) & (map->cap-1));
 			while(map->slot[empty].key != NULL) {
 				empty = (empty + 1) & (map->cap-1);
 			}
@@ -310,7 +310,7 @@ _pbcM_sp_rehash(struct map_sp *map) {
 	size_t size = map->size;
 	map->size = 0;
 	map->cap *= 2;
-	map->slot = (struct _pbcM_sp_slot *)HMALLOC(sizeof(struct _pbcM_sp_slot)*map->cap);
+	map->slot = (struct _pbcM_sp_slot *)HMALLOC((int)(sizeof(struct _pbcM_sp_slot)*map->cap));
 	memset(map->slot,0,sizeof(struct _pbcM_sp_slot)*map->cap);
 	size_t i;
 	for (i=0;i<size;i++) {
@@ -345,7 +345,7 @@ _pbcM_sp_query_insert_hash(struct map_sp *map, const char *key, size_t hash_full
 		if (map->cap <= map->size) 
 			goto _rehash;
 
-		int empty = (hash + 1) & (map->cap-1);
+		int empty = (int)((hash + 1) & (map->cap-1));
 		while(map->slot[empty].key != NULL) {
 			empty = (empty + 1) & (map->cap-1);
 		}
@@ -426,7 +426,7 @@ _find_first(struct map_sp *map)
 	size_t i;
 	for (i=0;i<map->cap;i++) {
 		if (map->slot[i].pointer) {
-			return i;
+			return (int)i;
 		}
 	}
 	return -1;
@@ -443,7 +443,7 @@ _find_next(struct map_sp *map, const char *key)
 		return -1;
 	for (;;) {
 		if (slot->hash == hash_full && strcmp(slot->key, key) == 0) {
-			int i = slot - map->slot + 1;
+			int i = (int)(slot - map->slot + 1);
 			while(i<map->cap) {
 				if (map->slot[i].pointer) {
 					return i;
